pyproject.toml
[tool.poetry]
name = "newsrag"
version = "0.1.0"
description = "A library and demo app for ingesting news feeds through advanced NLP pipelines."
authors = ["mattlbeck <17108752+mattlbeck@users.noreply.github.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
haystack-ai = "^2.7.0"
ollama-haystack = "^2.0.0"
sentence-transformers = "^3.3.1"
torch = "^2.5.1"
lxml-html-clean = "^0.4.1"
feedparser = "^6.0.11"
arrow = "^1.3.0"
top2vec = "^1.0.36"
gradio = "5.5.0"
llvmlite = "^0.43.0"
pytest-recording = "^0.13.2"
scipy = "<1.13"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ipykernel = "^6.29.5"


[tool.poetry.group.experiments.dependencies]
dvc = "^3.57.0"
jsonlines = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
